Replaced Timer by a scheduler
Added a scheduler in a separate thread to replace the timer functions. This allows messages to process a lot faster (removing events from the scheduler is quicker than cancelling timer threads) and reduces the cpu load since we don't have to spawn two new threads each time a message is received.